feat(compile): add `tag` transformer
The `tag` transformer inserts a global `Symbol` into the argument to
`tag`. The key for this `Symbol` is taken from a combination of the
`moduleMap` in the `tsconfig.json`, as well as the name of the type
supplied to the type argument of `tag`. This *should* prevent any
trouble that may arise from the assumptions bundlers make about ES
modules.