feat(observable): implement `Observable` and related constructs
Much of this package is adapted from
[ReactiveX/rxjs](http://github.com/ReactiveX/rxjs) with one major
difference. This package differentiates fatal errors (`defects`) from
typed failures. `Defects` are untyped (or, typed as `unknown`) and can
occur when a value is thrown, or when the `defect` constructor is used.
Returned errors, on the other hand, occur through use of the `fail`
constructor. `Defects`, like `error` in `rxjs`, halt the continuation of
an `Observable`; typed failures do not. This allows `Observers` to
process a stream of both values and typed failures, increasing their
error-handling potential for non-fatal errors.